3608: Refactoring namespace & unused references
