+++ src/NHSD.BuyingCatalogue.Solutions.API/Controllers/NativeMobile/MemoryAndStorageController.cs
@@ -9,7 +9,7 @@ using NHSD.BuyingCatalogue.Solutions.API.ViewModels.NativeMobile;
 using NHSD.BuyingCatalogue.Solutions.Application.Commands.UpdateSolutionMobileMemoryAndStorage;
 using NHSD.BuyingCatalogue.Solutions.Contracts.Queries;
 
-namespace NHSD.BuyingCatalogue.Solutions.API.Controllers
+namespace NHSD.BuyingCatalogue.Solutions.API.Controllers.NativeMobile
 {
     [Route("api/v1/solutions")]
     [ApiController]
