+++ tests/NHSD.BuyingCatalogue.API.IntegrationTests/Steps/BrowserHardwareRequirements/GetBrowserHardwareRequirementSteps.cs
@@ -1,6 +1,7 @@
 using System.Threading.Tasks;
 using FluentAssertions;
 using Newtonsoft.Json.Linq;
+using NHSD.BuyingCatalogue.API.IntegrationTests.Steps.Common;
 using TechTalk.SpecFlow;
 
 namespace NHSD.BuyingCatalogue.API.IntegrationTests.Steps.BrowserHardwareRequirements
