+++ tests/NHSD.BuyingCatalogue.API.IntegrationTests/Steps/BrowsersSupported/GetSupportedBrowserSteps.cs
@@ -1,9 +1,10 @@
 using System.Threading.Tasks;
 using FluentAssertions;
 using Newtonsoft.Json.Linq;
+using NHSD.BuyingCatalogue.API.IntegrationTests.Steps.Common;
 using TechTalk.SpecFlow;
 
-namespace NHSD.BuyingCatalogue.API.IntegrationTests.Steps
+namespace NHSD.BuyingCatalogue.API.IntegrationTests.Steps.BrowsersSupported
 {
     [Binding]
     internal sealed class GetSupportedBrowserSteps
