+++ tests/NHSD.BuyingCatalogue.API.IntegrationTests/Steps/ClientApplicationTypes/NativeMobileSectionSteps.cs
@@ -2,6 +2,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using FluentAssertions;
+using NHSD.BuyingCatalogue.API.IntegrationTests.Steps.Common;
 using TechTalk.SpecFlow;
 using TechTalk.SpecFlow.Assist;
