+++ tests/NHSD.BuyingCatalogue.API.IntegrationTests/Steps/ContactDetails/ContactDetailsGetSteps.cs
@@ -2,6 +2,7 @@ using System.Globalization;
 using System.Linq;
 using System.Threading.Tasks;
 using FluentAssertions;
+using NHSD.BuyingCatalogue.API.IntegrationTests.Steps.Common;
 using NHSD.BuyingCatalogue.API.IntegrationTests.Support;
 using TechTalk.SpecFlow;
 using TechTalk.SpecFlow.Assist;
