+++ tests/NHSD.BuyingCatalogue.API.IntegrationTests/Steps/Entities/DatabaseSteps.cs
@@ -1,8 +1,9 @@
 using System.Threading.Tasks;
+using NHSD.BuyingCatalogue.API.IntegrationTests.Steps.Common;
 using NHSD.BuyingCatalogue.Testing.Data;
 using TechTalk.SpecFlow;
 
-namespace NHSD.BuyingCatalogue.API.IntegrationTests.Steps
+namespace NHSD.BuyingCatalogue.API.IntegrationTests.Steps.Entities
 {
     [Binding]
     internal class DatabaseSteps
