+++ tests/NHSD.BuyingCatalogue.API.IntegrationTests/Steps/Entities/OrganisationEntitySteps.cs
@@ -4,7 +4,7 @@ using NHSD.BuyingCatalogue.Testing.Data.EntityBuilders;
 using TechTalk.SpecFlow;
 using TechTalk.SpecFlow.Assist;
 
-namespace NHSD.BuyingCatalogue.API.IntegrationTests.Steps
+namespace NHSD.BuyingCatalogue.API.IntegrationTests.Steps.Entities
 {
     [Binding]
     public sealed class OrganisationEntitySteps
