+++ tests/NHSD.BuyingCatalogue.API.IntegrationTests/Steps/Features/FeaturesSectionValidationSteps.cs
@@ -1,13 +1,12 @@
 using System.Collections.Generic;
 using System.Globalization;
-using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using FluentAssertions;
+using NHSD.BuyingCatalogue.API.IntegrationTests.Steps.Common;
 using NHSD.BuyingCatalogue.API.IntegrationTests.Support;
 using TechTalk.SpecFlow;
 
-namespace NHSD.BuyingCatalogue.API.IntegrationTests.Steps
+namespace NHSD.BuyingCatalogue.API.IntegrationTests.Steps.Features
 {
     [Binding]
     internal sealed class FeaturesSectionValidationSteps
