+++ tests/NHSD.BuyingCatalogue.API.IntegrationTests/Steps/Solution/SolutionCapabilitySteps.cs
@@ -1,6 +1,7 @@
 using System.Linq;
 using System.Threading.Tasks;
 using FluentAssertions;
+using NHSD.BuyingCatalogue.API.IntegrationTests.Steps.Common;
 using TechTalk.SpecFlow;
 using TechTalk.SpecFlow.Assist;
