+++ tests/NHSD.BuyingCatalogue.API.IntegrationTests/Steps/SolutionDescription/GetSolutionDescriptionSteps.cs
@@ -1,8 +1,9 @@
 using System.Threading.Tasks;
 using FluentAssertions;
+using NHSD.BuyingCatalogue.API.IntegrationTests.Steps.Common;
 using TechTalk.SpecFlow;
 
-namespace NHSD.BuyingCatalogue.API.IntegrationTests.Steps
+namespace NHSD.BuyingCatalogue.API.IntegrationTests.Steps.SolutionDescription
 {
     [Binding]
     internal sealed class GetSolutionDescriptionSteps
