+++ tests/NHSD.BuyingCatalogue.API.IntegrationTests/Steps/SubmitForReview/SubmitForReviewSteps.cs
@@ -3,11 +3,12 @@ using System.Globalization;
 using System.Linq;
 using System.Threading.Tasks;
 using FluentAssertions;
+using NHSD.BuyingCatalogue.API.IntegrationTests.Steps.Common;
 using NHSD.BuyingCatalogue.API.IntegrationTests.Support;
 using TechTalk.SpecFlow;
 using TechTalk.SpecFlow.Assist;
 
-namespace NHSD.BuyingCatalogue.API.IntegrationTests.Steps
+namespace NHSD.BuyingCatalogue.API.IntegrationTests.Steps.SubmitForReview
 {
     [Binding]
     internal sealed class SubmitForReviewSteps
